--- concert-reservation-application/src/main/java/concert/application/concertschedule/application/dto/request/ConcertScheduleCreateRequest.java
@@ -1,21 +1,0 @@
-package concert.application.concertschedule.application.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-public class ConcertScheduleCreateRequest {
-
-    private String concertName;
-    private LocalDateTime dateTime;
-
-    @Builder
-    public ConcertScheduleCreateRequest(String concertName, LocalDateTime dateTime){
-        this.concertName = concertName;
-        this.dateTime = dateTime;
-    }
-}
