+++ concert-reservation-application/src/main/java/concert/application/concertschedule/presentation/request/ConcertScheduleRequest.java
@@ -1,0 +1,19 @@
+package concert.application.concertschedule.presentation.request;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ConcertScheduleRequest {
+
+    private String token;
+    private long concertId;
+
+    @Builder
+    public ConcertScheduleRequest(String token, long concertId){
+        this.token = token;
+        this.concertId = concertId;
+    }
+}
