+++ concert-reservation-application/src/main/java/concert/application/concertschedule/presentation/request/SeatNumbersRequest.java
@@ -1,0 +1,18 @@
+package concert.application.concertschedule.presentation.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class SeatNumbersRequest {
+
+    private String token;
+    private long concertScheduleId;
+
+    public SeatNumbersRequest(String token, long concertScheduleId){
+        this.token = token;
+        this.concertScheduleId = concertScheduleId;
+    }
+
+}
