--- concert-reservation-application/src/main/java/concert/application/concertscheduleseat/application/dto/response/SeatNumbersResponse.java
@@ -1,23 +1,0 @@
-package concert.application.concertscheduleseat.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class SeatNumbersResponse {
-
-    private final List<Long> availableSeatNumbers;
-
-    @Builder
-    public SeatNumbersResponse(List<Long> availableSeatNumbers){
-        this.availableSeatNumbers = availableSeatNumbers;
-    }
-
-    public static SeatNumbersResponse of(List<Long> availableSeatNumbers){
-        return SeatNumbersResponse.builder()
-                .availableSeatNumbers(availableSeatNumbers)
-                .build();
-    }
-}
